refactor(web): extract account deletion recovery gate
